requirements_.txt
# deployment/requirements.txt

# Core ML Libraries
torch>=1.13.0
torchvision>=0.14.0
transformers>=4.21.0
sentence-transformers>=2.2.2
pillow>=9.2.0
numpy>=1.21.0
pandas>=1.5.0

# Vector Database & RAG
chromadb>=0.4.0
langchain>=0.0.300
langchain-community>=0.0.10

# API Framework
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0

# Frontend
streamlit>=1.28.0
plotly>=5.15.0
matplotlib>=3.6.0
seaborn>=0.12.0

# MLOps & Monitoring
mlflow>=2.5.0
psutil>=5.9.0
prometheus-client>=0.17.0

# Data Processing
requests>=2.28.0
beautifulsoup4>=4.11.0
lxml>=4.9.0
openpyxl>=3.0.0

# Testing
pytest>=7.1.0
pytest-cov>=4.0.0
pytest-asyncio>=0.21.0
httpx>=0.24.0

# Code Quality
black>=23.0.0
flake8>=5.0.0
mypy>=1.0.0
isort>=5.12.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
click>=8.1.0
tqdm>=4.64.0
loguru>=0.7.0

# Production
gunicorn>=21.0.0
redis>=4.5.0
celery>=5.3.0